Split sync diagnostics modules
