+++ service-worker.js
@@ -174,6 +174,9 @@ const APP_SHELL = [
   '/js/sync-environment.js',
   '/js/sync-identity.js',
   '/js/sync-diagnostics.js',
+  '/js/sync-diagnostics-context.js',
+  '/js/sync-diagnostics-snapshot.js',
+  '/js/sync-diagnostics-text.js',
   '/js/sync-diagnose-actions-context.js',
   '/js/sync-diagnose-relay-actions.js',
   '/js/sync-diagnose-identity-actions.js',
